requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Image Processing & ML
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
pytesseract==0.3.10
# face-recognition==1.3.0
# dlib==19.24.2

mediapipe==0.10.9

# LiveKit
livekit-api==0.4.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
boto3==1.34.0
amazon-transcribe==0.6.2

